re-teach .q/.lq to get a displayname from userinfo for left users

+++ Izzy-Moonbot/Modules/QuotesModule.cs
@@ -1,6 +1,8 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Data;
 using System.Linq;
+using System.Reflection;
 using System.Threading.Tasks;
 using Discord;
 using Discord.Commands;
@@ -17,11 +19,32 @@ public class QuotesModule : ModuleBase<SocketCommandContext>
 {
     private readonly Config _config;
     private readonly QuoteService _quoteService;
+    private readonly Dictionary<ulong, User> _users;
 
-    public QuotesModule(Config config, QuoteService quoteService)
+    public QuotesModule(Config config, QuoteService quoteService, Dictionary<ulong, User> users)
     {
         _config = config;
         _quoteService = quoteService;
+        _users = users;
+    }
+
+    // Usually <@id> is the best display format, but _users contains the names for some user ids
+    // who left the server yet are still recorded in quotes, so we can do better than Discord
+    // by fetching those ancient user names.
+    private string DisplayUserId(ulong userId, IIzzyGuild guild)
+    {
+        var potentialUser = guild.GetUser(userId);
+
+        // Still in the server, so <@id> will resolve just fine
+        if (potentialUser != null)
+            return $"<@{userId}>";
+
+        if (_users.TryGetValue(userId, out var user))
+            // This is the case where we have a name that Discord probably doesn't, so use it
+            return user.Username;
+
+        // Never mind, we know nothing after all, just let them render as <@123456>
+        return $"<@{userId}>";
     }
 
     [Command("quote")]
@@ -51,7 +74,7 @@ public class QuotesModule : ModuleBase<SocketCommandContext>
         {
             var (randomUserId, index, quote) = _quoteService.GetRandomQuote();
 
-            await context.Channel.SendMessageAsync($"<@{randomUserId}> **`#{index + 1}`:** {quote}", allowedMentions: AllowedMentions.None);
+            await context.Channel.SendMessageAsync($"{DisplayUserId(randomUserId, defaultGuild)} **`#{index + 1}`:** {quote}", allowedMentions: AllowedMentions.None);
             return;
         }
 
@@ -66,7 +89,7 @@ public class QuotesModule : ModuleBase<SocketCommandContext>
         }
 
         ulong userId = _quoteService.AliasExists(search)
-            ? _quoteService.ProcessAlias(search, defaultGuild).Id
+            ? _quoteService.ProcessAlias(search, defaultGuild)
             : await DiscordHelper.GetUserIdFromPingOrIfOnlySearchResultAsync(search, context, true);
         if (userId == 0)
         {
@@ -84,7 +107,7 @@ public class QuotesModule : ModuleBase<SocketCommandContext>
                 return;
             }
 
-            await context.Channel.SendMessageAsync($"<@{userId}> **`#{result.Value.Item1 + 1}`:** {result.Value.Item2}", allowedMentions: AllowedMentions.None);
+            await context.Channel.SendMessageAsync($"{DisplayUserId(userId, defaultGuild)} **`#{result.Value.Item1 + 1}`:** {result.Value.Item2}", allowedMentions: AllowedMentions.None);
         }
         // Get specific quote from a specific user
         else
@@ -102,7 +125,7 @@ public class QuotesModule : ModuleBase<SocketCommandContext>
                 return;
             }
 
-            await context.Channel.SendMessageAsync($"<@{userId}> **`#{number.Value}`:** {quote}", allowedMentions: AllowedMentions.None);
+            await context.Channel.SendMessageAsync($"{DisplayUserId(userId, defaultGuild)} **`#{number.Value}`:** {quote}", allowedMentions: AllowedMentions.None);
         }
     }
 
@@ -144,7 +167,7 @@ public class QuotesModule : ModuleBase<SocketCommandContext>
 
         ulong userId;
         if (_quoteService.AliasExists(search))
-            userId = _quoteService.ProcessAlias(search, defaultGuild).Id;
+            userId = _quoteService.ProcessAlias(search, defaultGuild);
         else
             userId = await DiscordHelper.GetUserIdFromPingOrIfOnlySearchResultAsync(search, context, true);
 
@@ -163,7 +186,7 @@ public class QuotesModule : ModuleBase<SocketCommandContext>
 
         PaginationHelper.PaginateIfNeededAndSendMessage(
             context,
-            $"Here's all the quotes I have for <@{userId}>.",
+            $"Here's all the quotes I have for {DisplayUserId(userId, defaultGuild)}.",
             quotes.Select((quote, index) => $"{index + 1}: {quote}").ToArray(),
             $"Run `{_config.Prefix}quote <user> <number>` to get a specific quote.\n" +
             $"Run `{_config.Prefix}quote <user>` to get a random quote from that user.\n" +
@@ -232,7 +255,10 @@ public class QuotesModule : ModuleBase<SocketCommandContext>
         // Check for aliases
         if (_quoteService.AliasExists(user))
         {
-            var quoteUser = _quoteService.ProcessAlias(user, context.Guild);
+            var aliasUserId = _quoteService.ProcessAlias(user, context.Guild);
+            var quoteUser = context.Guild.GetUser(aliasUserId);
+            if (quoteUser == null)
+                throw new TargetException("The user this alias referenced to cannot be found.");
 
             var newAliasUserQuote = await _quoteService.AddQuote(quoteUser, content);
 
@@ -310,7 +336,10 @@ public class QuotesModule : ModuleBase<SocketCommandContext>
         // Check for aliases
         if (_quoteService.AliasExists(user))
         {
-            var quoteUser = _quoteService.ProcessAlias(user, context.Guild);
+            var aliasUserId = _quoteService.ProcessAlias(user, context.Guild);
+            var quoteUser = context.Guild.GetUser(aliasUserId);
+            if (quoteUser == null)
+                throw new TargetException("The user this alias referenced to cannot be found.");
 
             await _quoteService.RemoveQuote(quoteUser, number.Value - 1);
 
@@ -393,7 +422,10 @@ public class QuotesModule : ModuleBase<SocketCommandContext>
 
             if (_quoteService.AliasExists(alias))
             {
-                var user = _quoteService.ProcessAlias(alias, context.Guild);
+                var aliasUserId = _quoteService.ProcessAlias(alias, context.Guild);
+                var user = context.Guild?.GetUser(aliasUserId);
+                if (user == null)
+                    throw new TargetException("The user this alias referenced to cannot be found.");
 
                 await context.Channel.SendMessageAsync(
                     $"Quote alias **{alias}** maps to user **{user.Username}#{user.Discriminator}**.", allowedMentions: AllowedMentions.None);
